BREAKING(path/unstable): move unstable overload of `basename` to `unstable-basename`
