+++ path/posix/unstable_basename.ts
@@ -1,39 +1,32 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import {
-  assertArgs,
-  lastPathSegment,
-  stripSuffix,
-} from "../_common/basename.ts";
-import { stripTrailingSeparators } from "../_common/strip_trailing_separators.ts";
-import { isPosixPathSeparator } from "./_util.ts";
+import { basename as stableBasename } from "./basename.ts";
+import { fromFileUrl } from "./from_file_url.ts";
 
 /**
  * Return the last portion of a `path`.
  * Trailing directory separators are ignored, and optional suffix is removed.
  *
+ * @experimental **UNSTABLE**: New API, yet to be vetted.
+ *
  * @example Usage
  * ```ts
- * import { basename } from "@std/path/posix/basename";
+ * import { basename } from "@std/path/posix/unstable-basename";
  * import { assertEquals } from "@std/assert";
  *
  * assertEquals(basename("/home/user/Documents/"), "Documents");
  * assertEquals(basename("/home/user/Documents/image.png"), "image.png");
  * assertEquals(basename("/home/user/Documents/image.png", ".png"), "image");
+ * assertEquals(basename(new URL("file:///home/user/Documents/image.png")), "image.png");
+ * assertEquals(basename(new URL("file:///home/user/Documents/image.png"), ".png"), "image");
  * ```
  *
  * @param path The path to extract the name from.
  * @param suffix The suffix to remove from extracted name.
  * @returns The extracted name.
  */
-export function basename(path: string, suffix?: string): string {
-  assertArgs(path, suffix ?? "");
-
-  const lastSegment = lastPathSegment(path, isPosixPathSeparator);
-  const strippedSegment = stripTrailingSeparators(
-    lastSegment,
-    isPosixPathSeparator,
-  );
-  return suffix ? stripSuffix(strippedSegment, suffix) : strippedSegment;
+export function basename(path: string | URL, suffix?: string): string {
+  path = path instanceof URL ? fromFileUrl(path) : path;
+  return stableBasename(path, suffix);
 }
