+++ path/unstable_basename.ts
@@ -2,24 +2,28 @@
 // This module is browser compatible.
 
 import { isWindows } from "./_os.ts";
-import { basename as posixBasename } from "./posix/basename.ts";
-import { basename as windowsBasename } from "./windows/basename.ts";
+import { basename as posixUnstableBasename } from "./posix/unstable_basename.ts";
+import { basename as windowsUnstableBasename } from "./windows/unstable_basename.ts";
 
 /**
  * Return the last portion of a path.
+ *
+ * @experimental **UNSTABLE**: New API, yet to be vetted.
  *
  * The trailing directory separators are ignored, and optional suffix is
  * removed.
  *
  * @example Usage
  * ```ts
- * import { basename } from "@std/path/basename";
+ * import { basename } from "@std/path/unstable-basename";
  * import { assertEquals } from "@std/assert";
  *
  * if (Deno.build.os === "windows") {
  *   assertEquals(basename("C:\\user\\Documents\\image.png"), "image.png");
+ *   assertEquals(basename(new URL("file:///C:/user/Documents/image.png")), "image.png");
  * } else {
  *   assertEquals(basename("/home/user/Documents/image.png"), "image.png");
+ *   assertEquals(basename(new URL("file:///home/user/Documents/image.png")), "image.png");
  * }
  * ```
  *
@@ -28,8 +32,8 @@ import { basename as windowsBasename } from "./windows/basename.ts";
  *
  * @returns The basename of the path.
  */
-export function basename(path: string, suffix?: string): string {
+export function basename(path: string | URL, suffix?: string): string {
   return isWindows
-    ? windowsBasename(path, suffix)
-    : posixBasename(path, suffix);
+    ? windowsUnstableBasename(path, suffix)
+    : posixUnstableBasename(path, suffix);
 }
