+++ path/windows/basename.ts
@@ -9,7 +9,6 @@ import {
 import { CHAR_COLON } from "../_common/constants.ts";
 import { stripTrailingSeparators } from "../_common/strip_trailing_separators.ts";
 import { isPathSeparator, isWindowsDeviceRoot } from "./_util.ts";
-import { fromFileUrl } from "./from_file_url.ts";
 
 /**
  * Return the last portion of a `path`.
@@ -29,33 +28,8 @@ import { fromFileUrl } from "./from_file_url.ts";
  * @param suffix The suffix to remove from extracted name.
  * @returns The extracted name.
  */
-export function basename(path: string, suffix?: string): string;
-/**
- * Return the last portion of a `path`.
- * Trailing directory separators are ignored, and optional suffix is removed.
- *
- * @experimental **UNSTABLE**: New API, yet to be vetted.
- *
- * @example Usage
- * ```ts
- * import { basename } from "@std/path/windows/basename";
- * import { assertEquals } from "@std/assert";
- *
- * assertEquals(basename("C:\\user\\Documents\\"), "Documents");
- * assertEquals(basename("C:\\user\\Documents\\image.png"), "image.png");
- * assertEquals(basename("C:\\user\\Documents\\image.png", ".png"), "image");
- * assertEquals(basename(new URL("file:///C:/user/Documents/image.png")), "image.png");
- * assertEquals(basename(new URL("file:///C:/user/Documents/image.png"), ".png"), "image");
- * ```
- *
- * @param path The path to extract the name from.
- * @param suffix The suffix to remove from extracted name.
- * @returns The extracted name.
- */
-export function basename(path: string | URL, suffix?: string): string;
-export function basename(path: string | URL, suffix = ""): string {
-  path = path instanceof URL ? fromFileUrl(path) : path;
-  assertArgs(path, suffix);
+export function basename(path: string, suffix?: string): string {
+  assertArgs(path, suffix ?? "");
 
   // Check for a drive letter prefix so as not to mistake the following
   // path separator as an extra separator at the end of the path that can be
